Add basic framebuffer implementation.
I've really made a mess of the code here, but things *sorta* work.
Graphics are being displayed but they're off-color and potentially
out of phase, too.  This needs more work (and a real big cleanup/refactor),
but it *is* starting to come together (^'.')^

partial commit: add barebones skeletons for framebuffer, shader