OpenGL: Bypass the render-to-texture FBO stuff and render directly to OpenGL
This is an experiment I did while debugging an FBO-related issue that occurs on
one computer that I own, but not on my main development machine.  I'd like to
hold on to this for now, but I have no intention of actually making it permanent.

The issue I'm trying to debug is that FBO depth buffers don't work on this
machine for some reason; the experiment here is to not use the FBO.  The result
is that outside of the context of an FBO, depth buffers do work (as does
everything else needed to render the Dreamcast's firmware menu).

The computer I'm seeing this issue on is a Debian machine with an Intel GPU.
The computer I'm not seeing this issue on is a Gentoo machine with an Nvidia
GPU.

TEMPORARY COMMIT; DO NOT MERGE TO MASTER