committing some stuff I've had sitting around in my git stash for a while
this makes the dual-issue pipeline faster by hard-coding the second issue.
I don't know if I can successfully put this on top of the other optimizations
I've been making recentiy, but this actually boosts performance by far more
than I would expect it to (12% to 15%!) so I want to keep it around for study.

I can't remember if there are any discrepancies in the cycle-counting caused by
this commit; if there are then that might explain why this commit seems so fast.