PARTIAL COMMIT: JIT: move registers into guest-independent slots for operations
This is a pre-requisite for a real register allocation system.  Eventually, I'm
going to be porting most of the SH4 instructions over to the native jit backend
so I don't have to lean on JIT_OP_FALLBACK.  When this happens, I want to be
able to store SH4 registers in native x86_64 registers so that the emitted code
doesn't keep needlessly going back to memory.

This is also eventually going to make the JIT's IL completely guest-independent,
which is important because one of these days I'm finally going to get around to
implementing the ARM7.